requirements.txt
python-dotenv
langchain
langchain-core
langchain-openai
langchain-community
langchain-chroma
chromadb
pypdf
langgraph
langchain-ollama
chainlit
ollama<0.4.0